+++ R/detailed_itineraries.R
@@ -4,8 +4,9 @@
 #'              or many origin destination pairs.
 #'
 #' @param r5r_core rJava object to connect with R5 routing engine
-#' @param origins,destinations either a spatial sf POINT object or a data.frame
-#'                            containing the columns 'id', 'lon', 'lat'
+#' @param origins,destinations a spatial sf POINT object with WGS84 CRS, or a
+#'                             data.frame containing the columns 'id', 'lon',
+#'                             'lat'.
 #' @param mode string. Transport modes allowed for the trips. Defaults to
 #'             "WALK". See details for other options.
 #' @param mode_egress string. Transport mode used after egress from public
@@ -14,13 +15,32 @@
 #'                    used.
 #' @param departure_datetime POSIXct object. If working with public transport
 #'                           networks, please check \code{calendar.txt} within
-#'                           the GTFS file for valid dates.
-#' @param max_walk_dist numeric. Maximum walking distance (in meters) for the
-#'                      whole trip. Defaults to no restrictions on walking, as
-#'                      long as \code{max_trip_duration} is respected.
-#' @param max_bike_dist numeric. Maximum cycling distance (in meters) for the
-#'                      whole trip. Defaults to no restrictions on cycling, as
-#'                      long as \code{max_trip_duration} is respected.
+#'                           the GTFS file for valid dates. See details for
+#'                           further information on how datetimes are parsed.
+#' @param max_walk_dist numeric. Maximum walking distance (in meters) to access
+#'                      and egress the transit network, or to make transfers
+#'                      within the network. Defaults to no restrictions as long
+#'                      as `max_trip_duration` is respected. The max distance is
+#'                      considered separately for each leg (e.g. if you set
+#'                      `max_walk_dist` to 1000, you could potentially walk up
+#'                      to 1 km to reach transit, and up to _another_ 1 km to
+#'                      reach the destination after leaving transit). Obs: if you
+#'                      want to set the maximum walking distance considering
+#'                      walking-only trips you have to set the `max_trip_duration`
+#'                      accordingly (e.g. to set a distance of 1 km assuming a
+#'                      walking speed of 3.6 km/h you have to set `max_trip_duration = 1 / 3.6 * 60`).
+#' @param max_bike_dist numeric. Maximum cycling distance (in meters) to access
+#'                      and egress the transit network. Defaults to no
+#'                      restrictions as long as `max_trip_duration` is respected.
+#'                      The max distance is considered separately for each leg
+#'                      (e.g. if you set `max_bike_dist` to 1000, you could
+#'                      potentially cycle up to 1 km to reach transit, and up
+#'                      to _another_  1 km to reach the destination after leaving
+#'                      transit). Obs: if you want to set the maximum cycling
+#'                      distance considering cycling-only trips you have to set
+#'                      the `max_trip_duration` accordingly (e.g. to set a
+#'                      distance of 5 km assuming a cycling speed of 12 km/h you
+#'                      have to set `max_trip_duration = 5 / 12 * 60`).
 #' @param max_trip_duration numeric. Maximum trip duration in minutes. Defaults
 #'                          to 120 minutes (2 hours).
 #' @param walk_speed numeric. Average walk speed in km/h. Defaults to 3.6 km/h.
@@ -38,9 +58,12 @@
 #' @param n_threads numeric. The number of threads to use in parallel computing.
 #'                  Defaults to use all available threads (Inf).
 #' @param verbose logical. `TRUE` to show detailed output messages (the default).
-#'                If verbose is set to `FALSE`, r5r prints a progress counter and
-#'                eventual `ERROR` messages. Setting `verbose` to  `FALSE` imposes
-#'                a small penalty for computation efficiency.
+#' @param progress logical. `TRUE` to show a progress counter. Only works when
+#'                `verbose` is set to `FALSE`, so the progress counter does not
+#'                interfere with R5's output messages. Setting `progress` to `TRUE`
+#'                may impose a small penalty for computation efficiency, because
+#'                the progress counter must be synchronized among all active
+#'                threads.
 #' @param drop_geometry logical. Indicates whether R5 should drop segment's
 #'                      geometry column. It can be helpful for saving memory.
 #'
@@ -64,8 +87,9 @@
 #' or 3.
 #'
 #' The default methodology for assigning LTS values to network edges is based on
-#' commonly tagged attributes of OSM ways. See more info about LTS at
-#' \url{https://docs.conveyal.com/learn-more/traffic-stress}. In summary:
+#' commonly tagged attributes of OSM ways. See more info about LTS in the original
+#' documentation of R5 from Conveyal at \url{https://docs.conveyal.com/learn-more/traffic-stress}.
+#' In summary:
 #'
 #'- **LTS 1**: Tolerable for children. This includes low-speed, low-volume streets,
 #'  as well as those with separated bicycle facilities (such as parking-protected
@@ -78,6 +102,9 @@
 #'- **LTS 4**: Tolerable for only “strong and fearless” cyclists. This includes streets
 #'  where cyclists are required to mix with moderate- to high-speed vehicular traffic.
 #'
+#'  For advanced users, you can provide custom LTS values by adding a tag
+#'  <key = "lts> to the `osm.pbf` file
+#'
 #' # Routing algorithm:
 #'  The detailed_itineraries function uses an R5-specific extension to the
 #'  McRAPTOR routing algorithm to find paths that are optimal or less than
@@ -87,6 +114,18 @@
 #'  McRAPTOR can be found in Delling et al (2015).
 #'  - Delling, D., Pajor, T., & Werneck, R. F. (2015). Round-based public transit
 #'   routing. Transportation Science, 49(3), 591-604.
+#'
+#' # Datetime parsing
+#'
+#' `r5r` ignores the timezone attribute of datetime objects when parsing dates
+#' and times, using the study area's timezone instead. For example, let's say
+#' you are running some calculations using Rio de Janeiro, Brazil, as your study
+#' area. The datetime `as.POSIXct("13-05-2019 14:00:00",
+#' format = "%d-%m-%Y %H:%M:%S")` will be parsed as May 13th, 2019, 14:00h in
+#' Rio's local time, as expected. But `as.POSIXct("13-05-2019 14:00:00",
+#' format = "%d-%m-%Y %H:%M:%S", tz = "Europe/Paris")` will also be parsed as
+#' the exact same date and time in Rio's local time, perhaps surprisingly,
+#' ignoring the timezone attribute.
 #'
 #' @return A LINESTRING sf with detailed information about the itineraries
 #'         between specified origins and destinations. Distances are in meters
@@ -99,7 +138,7 @@
 #'
 #' # build transport network
 #' data_path <- system.file("extdata/poa", package = "r5r")
-#' r5r_core <- setup_r5(data_path = data_path)
+#' r5r_core <- setup_r5(data_path = data_path, temp_dir = TRUE)
 #'
 #' # load origin/destination points
 #' points <- read.csv(file.path(data_path, "poa_points_of_interest.csv"))
@@ -135,6 +174,7 @@ detailed_itineraries <- function(r5r_core,
                                  shortest_path = TRUE,
                                  n_threads = Inf,
                                  verbose = TRUE,
+                                 progress = TRUE,
                                  drop_geometry = FALSE) {
 
 
@@ -242,6 +282,8 @@ detailed_itineraries <- function(r5r_core,
   # set verbose
   set_verbose(r5r_core, verbose)
 
+  # set progress
+  set_progress(r5r_core, progress)
 
   # call r5r_core method ----------------------------------------------------
 
@@ -276,7 +318,7 @@ detailed_itineraries <- function(r5r_core,
 
   } else {
 
-    path_options <- jdx::convertToR(path_options)
+    path_options <- java_to_dt(path_options)
 
     if (!is.data.frame(path_options)) {
 
@@ -288,10 +330,10 @@ detailed_itineraries <- function(r5r_core,
 
   }
 
+  if (nrow(path_options) == 0) return(path_options)
+
   # return either the fastest or multiple itineraries between an o-d pair (untie
   # it by the option number, if necessary)
-
-  data.table::setDT(path_options)
 
   path_options[, total_duration := sum(segment_duration, wait), by = .(fromId, toId, option)]
 
@@ -308,7 +350,7 @@ detailed_itineraries <- function(r5r_core,
     # itineraries with the same signature (sequence of routes) are filtered to
     # keep the one with the shortest duration
 
-    path_options[, temp_route := ifelse(route == "", mode, route)]
+    path_options[, temp_route := fifelse(route == "", mode, route)]
     path_options[, temp_sign := paste(temp_route, collapse = "_"), by = .(fromId, toId, option)]
 
     path_options <- path_options[path_options[, .I[total_duration == min(total_duration)],by = .(fromId, toId, temp_sign)]$V1]
@@ -323,7 +365,7 @@ detailed_itineraries <- function(r5r_core,
   # options
   path_options[, option := data.table::rleid(option), by = .(fromId, toId)]
 
-  # if results includes the geometry, convert path_options from data.frame to
+  # if results include the geometry, convert path_options from data.frame to
   # data.table with sfc column
   if (!drop_geometry) {