+++ R/street_network_to_sf.R
@@ -14,7 +14,7 @@
 #'
 #' # build transport network
 #' path <- system.file("extdata/poa", package = "r5r")
-#' r5r_core <- setup_r5(data_path = path)
+#' r5r_core <- setup_r5(data_path = path, temp_dir = TRUE)
 #'
 #' # extract street network from r5r_core
 #' street_net <- street_network_to_sf(r5r_core)
@@ -33,13 +33,13 @@ street_network_to_sf <- function(r5r_core) {
   network <- r5r_core$getStreetNetwork()
 
   # Convert vertices to SF (point)
-  vertices_df <- jdx::convertToR(network$get(0L), array.order = "column-major")
+  vertices_df <- java_to_dt(network$get(0L))
   vertices_sf <- sfheaders::sf_point(vertices_df, x='lon', y='lat', keep = TRUE)
   sf::st_crs(vertices_sf) <- 4326 # WGS 84
 
   # Convert edges to SF (linestring)
-  edges_df <- jdx::convertToR(network$get(1L), array.order = "column-major")
-  data.table::setDT(edges_df)[, geometry := sf::st_as_sfc(geometry)]
+  edges_df <- java_to_dt(network$get(1L))
+  edges_df[, geometry := sf::st_as_sfc(geometry)]
   edges_sf <- sf::st_sf(edges_df, crs = 4326) # WGS 84
 
   # gather in a list