+++ tests/testthat/test-setup_r5.R
@@ -1,5 +1,6 @@
 context("setup_r5")
 
+# skips tests on CRAN since they require a specific version of java
 testthat::skip_on_cran()
 
 path <- system.file("extdata/poa", package = "r5r")
@@ -7,17 +8,17 @@ path <- system.file("extdata/poa", package = "r5r")
 # expected behavior
 test_that("setup_r5 - expected behavior", {
 
-  testthat::expect_message( setup_r5(data_path = path, verbose = F) )
-  testthat::expect_message( setup_r5(data_path = path, verbose = T) )
+  testthat::expect_message( setup_r5(data_path = path, verbose = F, temp_dir = TRUE) )
+  testthat::expect_message( setup_r5(data_path = path, verbose = T, temp_dir = TRUE) )
 
-  testthat::expect_message( setup_r5(data_path = path, use_elevation=T) )
+  testthat::expect_message( setup_r5(data_path = path, use_elevation=T, temp_dir = TRUE) )
 
   # remove files GTFS
   #  file.rename(file.path(path, "poa.zip"), file.path(path, "poa.x"))
   # testthat::expect_message( setup_r5(data_path = path, verbose = F) )
   #  file.rename(file.path(path, "poa.x"), file.path(path, "poa.zip"))
 
-  testthat::expect_message(setup_r5(data_path = path, version='6.0.1', verbose = F))
+  testthat::expect_message(setup_r5(data_path = path, version='6.4.0', verbose = F, temp_dir = TRUE))
 
 })
 
@@ -44,5 +45,23 @@ test_that("setup_r5 - expected errors", {
   #   file.rename(file.path(path, "network2.x"), file.path(path, "network.dat"))
 
   })
+
+test_that("'overwrite' parameter works correctly", {
+
+  testthat::expect_error(setup_r5(path, overwrite = 1))
+
+  # since a network was already created, if overwrite = FALSE it should use it
+  testthat::expect_message(
+    r5r_core <- setup_r5(path, verbose = FALSE, temp_dir = TRUE),
+    regexp = "Using cached network\\.dat from "
+  )
+
+  # but if overwrite = TRUE, then it should create a new network anyway
+  testthat::expect_message(
+    r5r_core <- setup_r5(path, verbose = FALSE, overwrite = TRUE, temp_dir = TRUE),
+    regexp = "Finished building network\\.dat at "
+  )
+
+})
 
 stop_r5()